Bill S1004 version original - Peter Oberacker
BILL NUMBER: S1004 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the penal law, in relation to the causation of a fire or explosion during the course of committing or attempting to commit a controlled substance felony PURPOSE: To provide for criminal liability for Arson for a person who causes a fire or explosion during the course of committing a controlled substance felony. SUMMARY OF PROVISIONS: Section one would add a subdivision to Penal Law § 150.