Bill S980 version original - Mario Mattera
BILL NUMBER: S980 SPONSOR: MATTERA TITLE OF BILL: An act to amend the correction law, in relation to sex offender regis- tration of change of address and internet accounts PURPOSE: To provide clarity to organizations responsible for ensuring the infor- mation on the registry is up to date. SUMMARY OF PROVISIONS: Section 1.