Bill S973 version original - Mario Mattera
BILL NUMBER: S973 SPONSOR: MATTERA TITLE OF BILL: An act to amend the penal law, in relation to presumption of intent to sell PURPOSE: Provides that possession of 40 or more individual packages of heroin or a heroin mixture is presumptive evidence of a person's intent to sell it, thereby increasing the severity of the crime. SUMMARY: Section 1: Adds a new section 220.