Bill S976 version original - Mario Mattera
BILL NUMBER: S976 SPONSOR: MATTERA TITLE OF BILL: An act to amend the family court act, in relation to the basis for removal of children and the creation of "Melinda's Law" PURPOSE: To provide a safe harbor for a non-abusive spouse to report abuse while still retaining custody of children. SUMMARY: Section 1 of this bill provides that this act will be known as "Melinda's Law."Section 2 of this bill provides for the legislative intent.