Bill S939 version original - Jamaal Bailey
BILL NUMBER: S939 SPONSOR: BAILEY TITLE OF BILL: An act to amend the criminal procedure law, in relation to facilitating appellate review of rulings that implicate issues of public concern PURPOSE OF BILL: The purpose of the Act is to require the court to review a denial of a motion to suppress evidence on appeal. By ensuring that a court review such denials, the legislative intent of section 710.