Bill S910 version original - Robert Ortt
BILL NUMBER: S910 SPONSOR: ORTT TITLE OF BILL: An act to amend the mental hygiene law, in relation to emergency inter- vention for persons impaired by substances PURPOSE: This legislation seeks to create a robust involuntary substance abuse treatment procedure for individuals who are a danger to themselves and others as a result of their systematic abuse of drugs. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill removes substances from Section 22.09 of the mental hygiene law.