Bill S1043A version original - Toby Ann Stavisky
BILL NUMBER: S1043 Revised 1/10 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the education law, in relation to the dispensing of self-administered hormonal contraceptives PURPOSE OR GENERAL IDEA OF BILL: To reduce the number of unintended pregnancies by increasing access to contraceptive drugs for women in New York State. SUMMARY OF SPECIFIC PROVISIONS: Section 1.