Bill S1059 version original - Rachel May
BILL NUMBER: S1059 SPONSOR: MAY TITLE OF BILL: An act to amend the executive law, in relation to code enforcement training programs for code enforcement personnel PURPOSE: The purpose of this bill is to allow flexible, online code enforcement training for code enforcement personnel. SUMMARY OF PROVISIONS: Subdivision 4 of section 376-a of the executive law is amended to provide for an online and flexible option for code enforcement training.