Bill S1032 version original - Robert Jackson
BILL NUMBER: S1032 SPONSOR: JACKSON TITLE OF BILL: An act to amend the arts and cultural affairs law, in relation to the Amistad Commission PURPOSE OR GENERAL IDEA OF BILL: This bill would move the Amistad commission from the Department of State to the New York State Education Department. SUMMARY OF PROVISIONS: Section 1 amends § 57.52 of the Arts and Cultural Affairs Law by allo- cating the Amistad Commission from the Department of State to the State Education Department.