Bill S986 version original - Andrew Gounardes
BILL NUMBER: S986 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the election law, in relation to establishing the New York state business contract database PURPOSE OF BILL: To establish a statewide "pay to play" database of those entities who have a business relationship with the state, and to limit political contributions from those entities. SUMMARY OF PROVISIONS: Section 1.