Bill S946 version original - Kevin Thomas
BILL NUMBER: S946 SPONSOR: THOMAS TITLE OF BILL: An act to amend the general business law, in relation to labeling requirements for electric assisted bicycles PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to increase consumer awareness that electric assisted bicycles may not be allowed on public highways, roads or side- walks. SUMMARY OF PROVISIONS: Section 1 of this bill amends the general business law by adding a new section 399-ggg.