Bill S1007 version original - Peter Oberacker
BILL NUMBER: S1007 SPONSOR: OBERACKER TITLE OF BILL: An act to establish a task force to study the effects of New York state opting out of daylight saving time PURPOSE: To convene a commission of the commissioners of health, agriculture and markets, and economic development, or their designees, to study the impact and effects of daylight saving time. SUMMARY OF PROVISIONS: Section 1 authorizes and directs the secretary of state to convene the commission.