Bill S905 version original - James Tedisco
BILL NUMBER: S905 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the criminal procedure law, in relation to setting bail for defendants who pose a threat to public safety PURPOSE: To allow greater judicial discretion in setting a securing order with respect to ' a defendant's prior felony convictions, failure to make an appearance in court, or subsequent arrests while awaiting trial. SUMMARY OF PROVISIONS: § 1: Amends subdivision 1 of section 510.