Bill S966 version original - Gustavo Rivera
BILL NUMBER: S966 SPONSOR: RIVERA TITLE OF BILL: An act to amend the elder law, in relation to program eligibility for plans comparable to Medicare part D PURPOSE: To allow individuals who have comparable coverage to a Medicare Part D plan to also be eligible for EPIC if they otherwise qualify. SUMMARY OF PROVISIONS: Section 1 of the bill amends section 242 of the elder law, which sets forth the eligibility criteria for the Elderly Pharmaceutical Insurance Coverage (EPIC) Program.