Bill S938 version original - Jamaal Bailey
BILL NUMBER: S938 SPONSOR: BAILEY TITLE OF BILL: An act to amend the criminal procedure law, in relation to enabling appellate review of the fairness and appropriateness of an imposed sentence PURPOSE OF BILL: The purpose of the Act is to require the court to modify sentences that are determined to be illegal or unduly harsh or severe. The proposed bill would secure the right of a defendant to seek out a remedy to unjust sentences.