Bill S969 version original - Jamaal Bailey
BILL NUMBER: S969 SPONSOR: BAILEY TITLE OF BILL: An act to amend the civil service law, in relation to appointments from eligible lists SUMMARY OF PROVISIONS: This bill amends subdivision 4 of section 23 and subdivisions 1 and 2 of section 56 of the civil service law, adding optional alternate list provisions to these sections. JUSTIFICATION: This legislation seeks to expand the pool of qualified candidates for police officer positions.