Bill S985 version original - Gustavo Rivera
BILL NUMBER: S985 SPONSOR: RIVERA TITLE OF BILL: An act to amend the public health law, in relation to the appointment of the state medicaid director PURPOSE: This bill provides that the Medicaid Director shall be appointed by the Governor. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends section 208-a of the public health law as it relates to the appointment of the state medicaid director.