Bill S994 version original - Brad Hoylman-Sigal
BILL NUMBER: S994 REVISED 5/5/2023 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the civil practice law and rules, in relation to admis- sibility of evidence of a victim's sexual conduct, sexual predisposi- tion, or manner of dress PURPOSE OR GENERAL IDEA OF BILL: To extend the rape shield law to civil practice law. SUMMARY OF SPECIFIC PROVISIONS: Section 1. The civil practice law and rules is amended by adding a new section 4550 to read as follows: § 4550.