Bill S996 version original - Brad Hoylman-Sigal
BILL NUMBER: S996 SPONSOR: HOYLMAN TITLE OF BILL: An act to amend the education law, in relation to requiring public schools to offer plant-based food options in food service PURPOSE: This bill would require public schools to offer plant-based food options to a student upon request by a student or person in parental relation to the student. SUMMARY OF PROVISIONS: Section 1 adds a new section 915-a to the Education Law to: 1.