Bill S1075 version original - George Borrello
BILL NUMBER: S1075 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the labor law, in relation to paid sick leave provided by employers with full-time employee equivalents PURPOSE: To amend state labor law mandated allocations of sick leave to reflect employees hours and duration of employment SUMMARY OF PROVISIONS: Section 1. Subdivision 1 of section 196-b of the labor law, as added by section 1 of part J of chapter 56 of the laws of 2020 is amended. Section 2.