Bill S1165 version original - Peter Oberacker
BILL NUMBER: S1165 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the penal law, in relation to elevating an assault of a code enforcement official to the class D felony of assault in the second degree PURPOSE: The bill elevates assault of a code enforcement official in the second degree to a class D felony. SUMMARY OF PROVISIONS: Section 1: Amends subdivision 3 of section 120.