Bill S1191 version original - Roxanne J. Persaud
BILL NUMBER: S1191 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the education law, in relation to requiring certain health care providers to disclose certain facts in regards to probation to current and new patients PURPOSE OR GENERAL IDEA OF BILL:.: The purpose of the bill is to require certain health care providers to disclose the fact that the provider is on probation to current and new patients. SUMMARY OF SPECIFIC PROVISIONS: Section 1.