Bill S1120 version original - George Borrello
BILL NUMBER: S1120 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the penal law, the criminal procedure law and the civil practice law and rules, in relation to the justified use of physical force PURPOSE OF BILL: Removes the duty to retreat in scenarios justified under article thir- ty-five of the penal law and adds protections for the defendant. SUMMARY OF PROVISIONS: Section 1 provides for the removal of the duty of retreat in places where the defender is lawfully permitted to be.