Bill S1166 version original - Peter Oberacker
BILL NUMBER: S1166 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the penal law, in relation to increasing the penalty for rape in the first degree PURPOSE: Amends the penal law to increase the penalty for rape in the first-de- gree from a class 13' felony to a class `A-II' felony. SUMMARY OF PROVISIONS: Section 1 amends the penal code to increase the penalty for rape. Section 2 states the effective date.