Bill S1181 version original - Pete Harckham
BILL NUMBER: S1181 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the executive law, in relation to designating March twenty-fifth as "Medal of Honor Day" PURPOSE: Designates March 25th to be commemorated as Medal of Honor Day in New York State. SUMMARY OF PROVISIONS: Section one of the bill amends subdivision 3 of section 168-a of the executive law to designate March 25th to be known as "Medal of Honor Day.