Bill S1204 version original - Kevin S. Parker
BILL NUMBER: S1204 SPONSOR: PARKER TITLE OF BILL: An act to amend the public service law, in relation to establishing discount programs for low-income senior citizens and individuals with a disability PURPOSE OR GENERAL IDEA OF BILL: To provide a discount for senior citizens and disabled individuals in the state of New York for such persons' utility bills. SUMMARY OF PROVISIONS: Section 1 provides for the utility discount program.