Bill S1092 version original - Jamaal Bailey
BILL NUMBER: S1092 SPONSOR: BAILEY TITLE OF BILL: An act to amend the penal law, in relation to re-defining physically helpless PURPOSE OR GENERAL IDEA OF BILL: To amend the penal relation to redefining the definition of "physical helplessness" SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill amends subdivision. 7 of section 130.00 of she penal lastò to say that "Physically.