Bill S1114A version original - George Borrello
BILL NUMBER: S1114 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the criminal procedure law, in relation to designating uniformed court officers in the town of Busti, county of Chautauqua, as peace officers PURPOSE OR GENERAL IDEA OF BILL: Designates the uniformed court officers of the Town of Busti, as peace officers. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends Section 2.