Bill S1138 version original - Robert Ortt
BILL NUMBER: S1138 SPONSOR: ORTT TITLE OF BILL: An act to amend chapter 658 of the laws of 1978 relating to incorpo- ration of the Lewiston No. 1 Volunteer and Exempt Firemen's Benevolent Association, in relation to its purpose and the use of foreign fire insurance premium taxes PURPOSE: To authorize the Lewiston Volunteer Firefighters" Benevolent Association No.1 to provide the relief, aid and assistance of members and their families who are injured.