Bill S1097 version original - Jamaal Bailey
BILL NUMBER: S1097 SPONSOR: BAILEY TITLE OF BILL: An act to amend the general construction law, in relation to the desig- nation of Harriet Tubman day PURPOSE OR GENERAL IDEA OF BILL: The primary purpose of this legislation is to establish a public holiday that commemorates Harriet Tubman, her life achievements and her pivotal role in the 'Underground Railroad". The legislation will designate the tenth of March of each year, to be known as Harriet Tubman Day.