Bill S1153 version original - James Sanders Jr.
BILL NUMBER: S1153 SPONSOR: SANDERS TITLE OF BILL: An act to amend the tax law, in relation to establishing a returning veterans tax credit for businesses that hire veterans and disabled veterans PURPOSE: To provide tax credits for businesses that hire veterans and disabled veterans. SUMMARY OF PROVISIONS: Section 1 amends section 210-B of the tax law to add a new subdivision 52 which provides for a returning veterans tax credit.