Bill S1115 version original - George Borrello
BILL NUMBER: S1115 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the civil service law and the executive law, in relation to eliminating the maximum age limitation for appointment as a police officer or as an environmental conservation officer PURPOSE OR GENERAL IDEA OF BILL: This bill eliminates the maximum age for taking the civil service exam- ination for appointment as a police officer. SUMMARY OF PROVISIONS: Sections 1-4.