Bill S1214 version original - Leroy Comrie
BILL NUMBER: S1214 SPONSOR: COMRIE TITLE OF BILL: An act to amend the public authorities law, in relation to special powers of the authority PURPOSE: Will allow riders travelling within city limits to pay a fare for commu- ter rails equal to that of NYC Transit subways and buses, and to provide a free transfer in the same manner. SUMMARY OF PROVISIONS: Amends section 1266 of the public authorities law by adding a new subdi- vision to accomplish the above stated purpose.