Bill S1096 version original - Jamaal Bailey
BILL NUMBER: S1096 SPONSOR: BAILEY TITLE OF BILL: An act to amend the public health law, in relation to establishing a healthy savings pilot program PURPOSE OR GENERAL IDEA OF BILL: Americans and New Yorkers are not consuming the minimum amount of fruits and vegetables recommended by the USDA, due to a variety of issues. The Healthy Savings programs seeks to incentivize consumption of these vital elements to a healthy life.