Bill S1130 version original - James Skoufis
BILL NUMBER: S1130 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the penal law, in relation to the crimes of commercial bribery and larceny; and to amend the criminal procedure law, in relation to jurisdiction for personal identifying information theft crimes PURPOSE: To amend the penal law to bring prosecutions of certain white collar crimes into the 21st century and keep up with changing realities. SUMMARY OF PROVISIONS: Sections 1 and 2: Amends section 180.03 and 180.