Bill S1068 version original - George Borrello
BILL NUMBER: S1068 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the not-for-profit corporation law, in relation to exempting the East Dunkirk Volunteer Fire Company, Inc. from the requirement that the percentage of non-resident fire department members not exceed forty-five percent of the membership PURPOSE: To authorize non-resident volunteer membership within the East Dunkirk Volunteer Fire Department.