Bill S1201C version original - Kevin S. Parker
BILL NUMBER: S1201 SPONSOR: PARKER TITLE OF BILL: An act to amend the public service law, in relation to establishing discount programs for veterans PURPOSE OR GENERAL IDEA OF BILL: To provide a discount for veterans in the state of New York for such persons' utility bills. SUMMARY OF PROVISIONS: Section 1 provides for the utility discount program.