Bill S1116 version original - George Borrello
BILL NUMBER: S1116 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the election law, in relation to prohibiting ballot harvesting PURPOSE OR GENERAL IDEA OF BILL: This bill would amend the election law, to prohibit ballot harvesting, and to provide criminal penalties for such acts. SUMMARY OF SPECIFIC PROVISIONS: This bill would add a new section 8-414 of the election law to prohibit ballot harvesting.