Bill S1237 version original - Jamaal Bailey
BILL NUMBER: S1237 SPONSOR: BAILEY TITLE OF BILL: An act to amend the penal law and the general business law, in relation to criminal mischief and securities fraud PURPOSE OR GENERAL IDEA OF BILL: To raise the existing monetary thresholds for Felony-Level Criminal Mischief and Securities Fraud. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill amends subdivision 2 of section 145.05 of the Penal Law to raise the monetary property damage threshold from ($250) to $1,000.