Bill S1186 version original - Roxanne J. Persaud
BILL NUMBER: S1186 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to school bus parking on city streets PURPOSE: Regulates the parking of school buses operated by or pursuant to a contract with the New York City Board of Education. SUMMARY OF PROVISIONS: Section 1. The administrative code of the City of New York is amended by adding a new section 19-610 to read as follows: 3 § 19-608 School bus parking on a city street.