Bill S1145 version original - Liz Krueger
BILL NUMBER: S1145 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the executive law, in relation to the appointment and reporting duties of the superintendent of the state police, the appoint- ment of employees of the state police and the approval of certain duties of the state police PURPOSE OR GENERAL IDEA OF BILL: Under this legislation, the Superintendent shall he appointed by the Governor, with the advice and consent of the. Senate, to a single seven year term.