Bill S1189 version original - Roxanne J. Persaud
BILL NUMBER: S1189 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the criminal procedure law, in relation to jury deliber- ations PURPOSE OR GENERAL IDEA OF BILL: the purpose of the bill is to amend the criminal procedure law, in relation to jury deliberation. SUMMARY OF SPECIFIC PROVISIONS: Section 1. This act shall be known and may be cited as the "Tiarah Poyau Act". § 2. Section 310.