Bill S1083 version original - Leroy Comrie
BILL NUMBER: S1083 SPONSOR: COMRIE TITLE OF BILL: An act to amend the education law, in relation to establishing Eid-al- Adha and Eid-al-Fitr as school holidays PURPOSE: To establish Eid-al-Adha and Rid al-Fitr as a school holiday as a school holiday. SUMMARY OF PROVISIONS: Section 1. The education law is amended by adding a new section 3204-a to read as follows: 3 § 3204-a. Eid-al-Adha and Eid al-Fitr as a school holiday.