Bill S1123 version original - Jeremy Cooney
BILL NUMBER: S1123 SPONSOR: COONEY TITLE OF BILL: An act to repeal subdivision 8 of section 1242 of the vehicle and traf- fic law relating to the operation of bicycles with electric assist PURPOSE OR GENERAL IDEA OF BILL: Section one repeals subdivision 8 of section 1242 of the vehicle and traffic law. Section two provides the effective date. SUMMARY OF PROVISIONS: Electric bikes ("e-bikes") have been increasingly adopted as an ecofriendly, cheap alternative to driving a car.