Bill S1160 version original - James Skoufis
BILL NUMBER: S1160 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the education law, in relation to the location of the annual school district elections PURPOSE: Relates to the location of the annual school district elections. SUMMARY OF PROVISIONS: Section 1: Amends Section 2002 of the education law to include any building owned or leased by the school district to be used for their annual meeting, election, board vote.