Bill S1199B version original - Kevin S. Parker
BILL NUMBER: S1199 SPONSOR: PARKER TITLE OF BILL: An act to amend the public service law, in relation to establishing minimum standards for payment plans for eligible customers PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to direct the Public Service Commission to set minimum stand for payment plans established by utility companies in the state. SUMMARY OF PROVISIONS: Section 1 amends the public service law.