Bill S1148 version original - Liz Krueger
BILL NUMBER: S1148 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the election law, in relation to designating early poll- ing places PURPOSE:. This bill provides that public schools shall not be designated as early voting polling sites. SUMMARY OF PROVISIONS: Section 1: Amends § 4-104 of the Election Law to provide that public buildings shall be designated as polling places to the extent practica- ble, however in no case shall a public school be designated as an early voting polling location.