Bill S1171 version original - Peter Oberacker
BILL NUMBER: S1171 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to eliminating additional fees for special license plates issued to veterans PURPOSE: To eliminate additional application fees and annual fees for special license plates issued to veterans of the armed forces of the United States. SUMMARY OF PROVISIONS: Section 1 amends subdivision 1-a of section 404 of the vehicle and traf- fic law, as amended by chapter 105 of the laws of 2017.