Bill S1177 version original - Pete Harckham
BILL NUMBER: S1177 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the insurance law, in relation to mandating commercial insurance coverage of peer support services as part of treatment for substance use disorder PURPOSE: This bill will mandate that commercial insurance coverage include peer support services as part of substance use disorder. SUMMARY OF PROVISIONS: Section 1 amends paragraph 31 of subsection (i) of section 3216 of the insurance law, as added by chapter 41 of the laws of 2014.