Bill S1142 version original - Robert Ortt
BILL NUMBER: S1142 SPONSOR: ORTT TITLE OF BILL: An act to amend the tax law, in relation to providing that excess investment tax credit amounts may be refundable to operators of a farm operation PURPOSE: To allow farmers to receive the investment tax credit as a fully refund- able credit. SUMMARY OF PROVISIONS: Section 1: Amends subdivision 1 of section 210-B of the tax law to provide that excess investment tax credit amounts may be refundable to operators of a farm operation.