Bill S1094 version original - Jamaal Bailey
BILL NUMBER: S1094 SPONSOR: BAILEY TITLE OF BILL: An act to amend the general obligations law, in relation to providing for compensation for injury or death caused by the negligent entrustment of a firearm PURPOSE: To extend the negligent entrustment liability of firearm manufacturers and sellers to the maximum extent allowable under federal law. SUMMARY OF PROVISIONS: Section 1 establishes the legislative intent.