Bill S1226 version original - Robert Ortt
BILL NUMBER: S1226 SPONSOR: ORTT TITLE OF BILL: An act to amend the military law, in relation to determining that disa- bility compensation or benefits received by veterans from the United States department of veterans affairs are not to be treated as income or resources for purposes of determining eligibility for means-tested programs PURPOSE: To ensure assistance and benefits received by veterans are not handled as income for capacity of means-tested programs. SUMMARY OF PROVISIONS: Section 1.