Bill S1184 version original - James Tedisco
BILL NUMBER: S1184 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the public service law, in relation to requiring legis- lative approval of increases in utility charges PURPOSE OR GENERAL IDEA OF BILL: To require in addition to the consent of the Public Service Commission, Legislative approval before implementing increases in fees, rentals and charges for the furnishing or rendition of gas or electric power. SUMMARY OF SPECIFIC PROVISIONS: Section 1.