Bill S1290 version original - Michelle Hinchey
BILL NUMBER: S1290 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the state law, in relation to a state flag for raising awareness of veteran suicide SUMMARY OF PROVISIONS: Section 1 adds a new section 92 to state law to establish the *Suicide Awareness and Remembrance Flag" as the official state flag for the rais- ing of awareness of veteran suicide. The commissioner of general services shall establish a protocol for the flying of such flag. Section 2 is the effective date.