Bill S1294 version original - Rachel May
BILL NUMBER: S1294 SPONSOR: MAY TITLE OF BILL: An act to amend the legislative law, in relation to the creation of the division of research and analysis to assist the legislature PURPOSE OR GENERAL IDEA OF BILL: This amends section 7-a of the legislative law to create a nonpartisan Division of Research and Analysis within the Legislative Library to provide research and analytical support for members of the legislature. SUMMARY OF SPECIFIC PROVISIONS: Section 1 lays out the legislative intent.