Bill S1427 version original - Jamaal Bailey
BILL NUMBER: S1427 SPONSOR: BAILEY TITLE OF BILL: An act creating a temporary state commission to examine the degree and effect of segregation in primary and secondary schools; and providing for the repeal of such provisions upon expiration thereof PURPOSE OR GENERAL IDEA OF BILL: Authorizes the state department of education to investigate. the degree and effect of segregation in flew York's primary and secondary schools, and make subsequent recommendations to the governor and legislature.