Bill S1435 version original - Joseph A. Griffo
BILL NUMBER: S1435 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the legislative law, in relation to requiring bills to embrace a single subject PURPOSE: To prevent the creation of legislation which encompasses multiple issue areas. SUMMARY OF PROVISIONS: Section 1 adds a new Section 49 to the legislative law. JUSTIFICATION: This bill is meant to prevent omnibus legislation from being presented to the Legislature.