Bill S1254 version original - Rachel May
BILL NUMBER: S1254 SPONSOR: MAY TITLE OF BILL: An act to amend the education law, in relation to requiring the commis- sioner of education to develop and implement an expanded curriculum or instruction on the history of the labor movement in the United States PURPOSE: This bill instructs the State Education Commissioner to develop and implement an expanded set of instructional or curriculum materials on the history of the labor movement in the United States. SUMMARY OF PROVISIONS: Section 1.