Bill S1434 version original - Joseph A. Griffo
BILL NUMBER: S1434 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the state finance law, in relation to requiring itemiza- tion, by school district, of proposed appropriations to the education department for the support of school districts PURPOSE: Requires the budget submitted by the governor to include an itemization, by each individual school district, of appropriations for the support of school districts. SUMMARY OF PROVISIONS: Section 1.