Bill S1292A version original - Kevin S. Parker
BILL NUMBER: S1292 SPONSOR: PARKER TITLE OF BILL: An act to amend the environmental conservation law, in relation to establishing the "clean fuel standard of 2023" PURPOSE: This legislation amends the environmental conservation law to require the development of a clean fuel standard in New York State to reduce greenhouse gas (GHG) emissions in the transportation sector to achieve the goals set forth in the climate leadership and community protection act. SUMMARY OF PROVISIONS: Section 1.