Bill S1251 version original - Patrick M. Gallivan
BILL NUMBER: S1251 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the penal law, in relation to aggravated harassment of an employee by an incarcerated individual PURPOSE OR GENERAL IDEA OF BILL: This legislation removes language as it relates to causing an employee to come into contact with certain body fluid by an incarcerate individ- ual. SUMMARY OF SPECIFIC PROVISIONS: Section one amends Section 240.