Bill S1398 version original - John Liu
BILL NUMBER: S1398 SPONSOR: LIU TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to repair or replacement of damaged residential lateral sewer and water pipes PURPOSE OR GENERAL IDEA OF BILL: Homeowners shall be reimbursed the full cost of any repairs that are the result damages from a third party. SUMMARY OF PROVISIONS: Section 1. The administrative code of the city of New York is amended by adding two new sections 24-316.