Bill S1387 version original - Leroy Comrie
BILL NUMBER: S1387 SPONSOR: COMRIE TITLE OF BILL: An act to amend the public service law, in relation to requiring cable companies to charge the same rate for the same service PURPOSE: To require cable companies to charge the same rate for the same service in the State of New York. SUMMARY OF PROVISIONS: Section 1: The Public Service law is amended to state no cable tele- vision company shall charge a different rate for the same service to different subscribers.