Bill S1337 version original - Leroy Comrie
BILL NUMBER: S1337 SPONSOR: COMRIE TITLE OF BILL: An act to amend the general municipal law and the public authorities law, in relation to notice procedures prior to approval of projects by industrial development agencies PURPOSE OF BILL: This is a chapter amendment that makes changes to provisions of L.2022, c.766. SUMMARY OF SPECIFIC PROVISIONS: This legislation amends the underlying chapter by permitting notices via email.