Bill S1298A version original - Leroy Comrie
BILL NUMBER: S1298 SPONSOR: COMRIE TITLE OF BILL: An act to amend the general business law, in relation to the use of voice recognition features PURPOSE: This bill would regulate voice recognition features in Internet-connect- ed devices SUMMARY OF PROVISIONS: Section 1 adds a new article 32-a to the general business law to regu- late voice recognition features. Subdivision 1 of section 676 sets forth definitions to be used in the section.