Bill S1313 version original - Robert Ortt
BILL NUMBER: S1313 SPONSOR: ORTT TITLE OF BILL: An act to amend the executive law, in relation to executive orders relating to cooperation with federal law enforcement PURPOSE: To block the governor from ordering state police to not comply with federal law enforcement. SUMMARY OF PROVISIONS: Section 1 adds a new Section 233 to the executive law to provide that the executive branch may not direct state police to not cooperate with federal law enforcement.