Bill S1407 version original - Robert Ortt
BILL NUMBER: S1407 SPONSOR: ORTT TITLE OF BILL: An act to amend the penal law, in relation to allowing rifles, shotguns, and firearms on college and university campuses PURPOSE: This section allows for the possession of a rifle, shotgun, or firearm on a college or university campus. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends section 265.