Bill S1268 version original - Robert Ortt
BILL NUMBER: S1268 SPONSOR: ORTT TITLE OF BILL: An act to amend the criminal procedure law, in relation to permitting the electronic appearance of a defendant in the county of Orleans PURPOSE: To amend the current Criminal Procedure allowing the electronic appear- ance of defendants in Orleans County. SUMMARY OF PROVISIONS: Subdivision 1 of Section 182.