Bill S1390 version original - James Tedisco
BILL NUMBER: S1390 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the executive law and the election law, in relation to prohibiting sex offenders who are on parole and in civil confinement from voting in certain elections PURPOSE OF THE BILL: To prohibit sex offenders who are on parole and in civil confinement from voting in certain elections SUMMARY OF PROVISIONS: Section 1. The executive law is amended by adding a new section 18-a to read as follows: § 18-a. Restriction on voting rights.