Bill S1302 version original - Leroy Comrie
BILL NUMBER: S1302 SPONSOR: COMRIE TITLE OF BILL: An act to amend the public housing law, in relation to directing the New York city housing authority to establish an occupant protection plan for each project constructed during or prior to 1978 PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to require the New York City Housing Author- ity to create an Occupant Protection Plan. SUMMARY OF PROVISIONS: Section one: Amends the public housing law to add a new section, 402d.